requirements.txt
streamlit==1.28.0
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.3.0
anthropic==0.7.7
google-generativeai==0.3.0
lxml==4.9.3
